requirements.txt
mysql-connector-python
python-dotenv
tabulate
plantuml 